+++ src/App.js
@@ -1,5 +1,5 @@
 import "./App.scss";
-import Header from "./components/header";
+import Header from "./components/header/header";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Footer from "./components/footer/footer";
